requirements.txt
flask==3.0.3
flask-security-too==5.4.2
flask-sqlalchemy==3.1.1
email-validator==2.2.0
passlib[bcrypt]==1.7.4
